Add benchmark and organize
