+++ Cargo.toml
@@ -27,6 +27,14 @@ rand = "0.8.5"
 globset = "0.4"
 time = { version = "0.3", features = ["local-offset", "formatting"]}
 
+[dev-dependencies]
+criterion = { version = "0.4", features = ["html_reports"] }
+
 [[bin]]
 name = "acs"
-path = "src/main.rs"
+path = "src/bin/main.rs"
+
+
+[[bench]]
+name = "system_benchmark"
+harness = false
