+++ src/daemon.rs
@@ -37,24 +37,24 @@ use efcl::{color, Color};
 use nix::unistd::Uid;
 use serde::Serialize;
 
-use super::config::Config;
-use super::cpu::{Speed, CPU};
-use super::gov::Gov;
-use super::graph::{Graph, GraphType, Grapher};
-use super::logger;
-use super::logger::Interface;
-use super::network::{hook, listen};
-use super::power::battery::{has_battery, Battery};
-use super::power::lid::{Lid, LidRetriever, LidState};
-use super::power::{Power, PowerRetriever};
-use super::settings::Settings;
-use super::setup::{inside_docker_message, inside_wsl_message};
-use super::system::{
+use crate::config::Config;
+use crate::cpu::{Speed, CPU};
+use crate::gov::Gov;
+use crate::graph::{Graph, GraphType, Grapher};
+use crate::logger;
+use crate::logger::Interface;
+use crate::network::{hook, listen};
+use crate::power::battery::{has_battery, Battery};
+use crate::power::lid::{Lid, LidRetriever, LidState};
+use crate::power::{Power, PowerRetriever};
+use crate::settings::Settings;
+use crate::setup::{inside_docker_message, inside_wsl_message};
+use crate::system::{
     check_available_governors, check_cpu_freq, check_cpu_temperature, check_cpu_usage,
     get_highest_temp, inside_docker, inside_wsl, list_cpus,
 };
-use super::terminal::terminal_width;
-use super::Error;
+use crate::terminal::terminal_width;
+use crate::error::Error;
 use crate::csv::{gen_writer, CSVWriter, Writer};
 use crate::display::{print_battery_status, print_turbo_status};
 use crate::proc::{parse_proc_file, read_proc_stat_file, ProcStat};
