+++ src/lib.rs
@@ -1,0 +1,21 @@
+pub mod args;
+pub mod config;
+pub mod cpu;
+pub mod csv;
+pub mod daemon;
+pub mod display;
+pub mod error;
+pub mod gov;
+pub mod graph;
+pub mod interactive;
+pub mod interface;
+pub mod logger;
+pub mod network;
+pub mod power;
+pub mod proc;
+pub mod settings;
+pub mod setup;
+pub mod sysfs;
+pub mod system;
+pub mod terminal;
+pub mod thermal;
